Engine Class

The Engine class is a core component of the oRm framework, responsible for managing database connections and providing methods for interacting with the database. It supports both direct connections and connection pooling, offering flexibility in how database resources are managed.

Key features:

Manages database connections (single or pooled)

Provides methods for executing SQL queries and commands

Allows creation of TableModel objects for ORM operations

Supports persistent connections for improved performance

Methods

Public methods

Engine$new()

Engine$get_connection()

Engine$close()

Engine$list_tables()

Engine$get_query()

Engine$execute()

Engine$model()

Engine$set_transaction_state()

Engine$get_transaction_state()

Engine$clone()

Method new()

Create an Engine object

Usage

Engine$new(..., conn_args = list(), use_pool = FALSE, persist = FALSE)

Arguments

...

Additional arguments to be passed to DBI::dbConnect

conn_args

A list of arguments to be passed to DBI::dbConnect

use_pool

Logical. Whether or not to make use of the pool package for connections to this engine

persist

Logical. Whether to keep the connection open after operations (default: FALSE) Get a connection to the database

Method get_connection()

Usage

Engine$get_connection()

Returns

A DBIConnection object or a pool object

Method close()

Close the database connection or pool

Usage

Engine$close()

Returns

NULL

Method list_tables()

List tables in the database connection

Usage

Engine$list_tables()

Returns

A character vector of table names Execute a SQL query and return the result as a data.frame

Method get_query()

Usage

Engine$get_query(sql)

Arguments

sql

SQL query

Returns

A data.frame Execute a SQL query and return the number of rows affected

Method execute()

Usage

Engine$execute(sql)

Method model()

Create a new TableModel object for the specified table

Usage

Engine$model(tablename, ..., .data = list())

Arguments

tablename

Name of the table

...

Additional arguments passed to the TableModel constructor

.data

A named list of the arguments for the TableModel constructor

Returns

A new TableModel object

Method set_transaction_state()

Usage

Engine$set_transaction_state(state)

Method get_transaction_state()

Usage

Engine$get_transaction_state()

Method clone()

The objects of this class are cloneable with this method.

Usage

Engine$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.